Rosette API text embedding service

Provides the embedding for the input text. Returns a single vector of floating point numbers for your input, known as a text embedding. Among other uses, a text embedding enables you to calculate the similarity between two documents or two words.

ros_embedding(content, genre = NULL, language = NULL)

Arguments

content

either a URI or character vector of content to process

genre

document genre (optional)

language

Language: ISO 639-3 code (optional)

Examples

ros_embedding("Cambridge, Massachusetts")

#> $embedding
#>   [1] -0.0752000000 -0.0404000020  0.0145000005 -0.0042333330  0.0155666670
#>   [6] -0.0446333330 -0.0434666700 -0.0389333330  0.0051999990  0.0075333330
#>  [11]  0.0176999980  0.0063666664  0.0995999950  0.0368666650  0.0041333330
#>  [16]  0.0114333330 -0.0718666600 -0.0056666670  0.0292999990 -0.0372333340
#>  [21]  0.0140666660  0.0185000000  0.0315000000  0.0328333300 -0.0391666700
#>  [26] -0.0260333340  0.0296666680 -0.0132333330  0.0316000020 -0.0243333340
#>  [31]  0.0080333330 -0.0493333340 -0.0105666680  0.0586666650  0.0247666690
#>  [36] -0.0022999998  0.0222000010  0.0470333300 -0.0057333335 -0.0166666680
#>  [41]  0.0265000020 -0.0437000000  0.0180999990  0.0042999997  0.0359333340
#>  [46] -0.0004000002  0.0404000000 -0.0460666680 -0.0234666670  0.0085999990
#>  [51] -0.0754666700  0.0490000000  0.0139333340  0.0101000000  0.0081000000
#>  [56]  0.0297666660  0.0409000000  0.0825000000  0.0261666680 -0.0144000010
#>  [61] -0.0320000000  0.0796333300 -0.0867333340 -0.0097000000  0.0608666650
#>  [66] -0.0304666660  0.0263666680  0.0243333340 -0.0053333340 -0.0320666700
#>  [71]  0.0164666680  0.0602333360 -0.0187000020 -0.0837999950 -0.0199666680
#>  [76]  0.0530000030  0.0288000010 -0.0034666670 -0.0295000000  0.0100333330
#>  [81]  0.0144666670 -0.0306999980 -0.0212666670  0.0173666640 -0.0106999995
#>  [86]  0.0117666675  0.0201666670 -0.0528666680  0.0360666660 -0.0354666670
#>  [91] -0.0425999950 -0.0200333330 -0.0325333330 -0.0044333330 -0.0050666668
#>  [96]  0.0176000000  0.0247000000  0.0460666700 -0.0066333340 -0.0360666660
#> [101]  0.0137333350  0.0283000000 -0.0006999998 -0.0697333300  0.0162666660
#> [106]  0.0235000010 -0.0237999990 -0.0310666650 -0.0028333322  0.0207333320
#> [111]  0.0380333330 -0.0260666660  0.0286333320  0.0087000010 -0.0021333330
#> [116]  0.0592666600  0.0016000023 -0.0206000000 -0.0154333340  0.0154999990
#> [121]  0.0380333330  0.0093000000  0.0343000000 -0.0230999980  0.0021000002
#> [126] -0.0248333350 -0.0186666680 -0.0824333350  0.0680666640  0.0232666660
#> [131] -0.0690333300 -0.0092000000 -0.0623666640  0.0494333360 -0.0100333330
#> [136] -0.0433333330 -0.0261000000 -0.0160333330 -0.0139999995 -0.0310333350
#> [141]  0.0076666675  0.0073333327  0.0294333340  0.0276666660  0.0356666700
#> [146] -0.0393666660 -0.0078333350  0.0045666660 -0.0346000000 -0.0100333320
#> [151]  0.0230000020  0.0045666670  0.0870000050 -0.0301000010 -0.0366333350
#> [156]  0.0036333340  0.0482999980  0.0340333320  0.0150666670 -0.0128666675
#> [161]  0.0485000000  0.0571666660  0.0472333320  0.0093333330  0.0089666670
#> [166]  0.0349333320  0.0028666675 -0.0154000000 -0.0276666660  0.0357666650
#> [171]  0.0497000030 -0.0076666675  0.0225333330 -0.0352666680  0.0245666650
#> [176]  0.0007333332  0.0663666650 -0.0158333330 -0.0274000000 -0.0187333320
#> [181] -0.0364333350 -0.0028999995  0.0432666650  0.0230666680 -0.0333333350
#> [186] -0.0327333320 -0.0349333320 -0.0563333330 -0.0128333340 -0.0362000000
#> [191] -0.0555000000  0.0114000000  0.0515999980 -0.0210333320 -0.0466666670
#> [196] -0.0224000010 -0.0042999997  0.0325333330 -0.0415000000 -0.0283666680
#> [201]  0.0169333310  0.0028999990  0.0178333320  0.0298333340  0.0023666670
#> [206]  0.0307000000 -0.0201333330 -0.0067666660  0.0023000005  0.0297000020
#> [211]  0.0078000003  0.0374333340 -0.0182666650  0.0041666660  0.0097999990
#> [216] -0.0441000000  0.0149000000 -0.0431000000 -0.0047999993 -0.0303333350
#> [221]  0.0162000000  0.0011666663  0.0543333330  0.0073333345  0.0195666680
#> [226] -0.0118666660 -0.0021666668  0.0211000000 -0.0072666670 -0.0009333330
#> [231]  0.0336000000 -0.0010666667  0.0396666680 -0.0533333300  0.0354333330
#> [236]  0.0334666630  0.0307666660 -0.0082000000  0.0375333300 -0.0389000000
#> [241]  0.0465666650 -0.0070999996  0.0249666660  0.0126666660  0.0144000010
#> [246]  0.0126333340 -0.0533666650 -0.0556666660  0.0034333344  0.0136666670
#> [251]  0.0031000000 -0.0010666671  0.0139999995 -0.0325333330  0.0318333320
#> [256]  0.0330333340 -0.0203666670  0.0062333330 -0.0298333340  0.0097666670
#> [261] -0.0196333340 -0.0065666675 -0.0191999990 -0.0246000000 -0.0030333325
#> [266]  0.0344000000  0.0068666670  0.0505000000  0.0015000001 -0.0190666670
#> [271] -0.0225666670  0.0099000000 -0.0008333332  0.0168000000  0.0352333340
#> [276]  0.0193333350  0.0214333330 -0.0194333330  0.0199000000 -0.0006666670
#> [281]  0.0078000003 -0.0300999970  0.0262666700  0.0511000000 -0.0056666670
#> [286] -0.0210666680 -0.0070000007 -0.0582333360 -0.0107333330  0.0442666670
#> [291] -0.0171333330 -0.0005999999  0.0088000010  0.0140333325 -0.0251666660
#> [296] -0.0054333340  0.0244666700  0.0111000010 -0.0223000000 -0.0383000000
#> [301]  0.0050333333 -0.0090000000 -0.0560000050 -0.0134333335 -0.0083333330
#> [306] -0.0321333330  0.0311666650  0.0243000000 -0.0279999990 -0.0245666650
#> [311]  0.0201333330 -0.0158666650  0.0065666675  0.0034000005  0.0130000000
#> [316] -0.0092666660  0.0438000000 -0.0515999980  0.0110333340 -0.0245333330
#> [321] -0.0003000014 -0.0017666664  0.0339666640  0.0147333330  0.0122333330
#> [326]  0.0163666660  0.0014333352  0.0094333340  0.0078000000 -0.0605666670
#> [331]  0.0047999998  0.0028333329  0.0002333332 -0.0616000000 -0.0309666640
#> [336]  0.0167333330  0.0472333320  0.0056999996 -0.0387333330  0.0305333340
#> [341] -0.0289666670  0.0040000000 -0.0112000000  0.0482333300 -0.0079333335
#> [346]  0.0149333340  0.0347666700  0.0337666700 -0.0221666680 -0.0377666650
#> [351]  0.0154666680  0.0490000000  0.0458333340 -0.0322000000 -0.0213666650
#> [356] -0.0131333330  0.0333333350 -0.0087999990 -0.0022000000  0.0570000000
#> [361] -0.0251666680  0.0314333330 -0.0162666660 -0.0088000000 -0.0081333340
#> [366] -0.0014999993 -0.0103666670  0.0213999990 -0.0129000010 -0.0165333340
#> [371] -0.0341000000  0.0142333330  0.0246666670  0.0153333330 -0.0374333340
#> [376]  0.0213000010 -0.0095333345 -0.0001666665  0.0005666666 -0.0440000000
#> [381]  0.0185666660 -0.0073333330 -0.0141000000 -0.0169333330 -0.0118999990
#> [386] -0.0130333330 -0.0028333336  0.0357666650 -0.0514333360  0.0583666670
#> [391]  0.0205333330  0.0053333340 -0.0223000000  0.0251000000  0.0281333330
#> [396] -0.0019666664  0.0176333340  0.0087000000 -0.0310333320  0.0052666664
#> [401] -0.0359000000  0.0340333320 -0.0006000002 -0.0292333350 -0.0246333330
#> [406] -0.0601666700  0.0214999990 -0.0663333300 -0.0121666670 -0.0222000010
#> [411]  0.0112333335 -0.0306666660  0.0230999980 -0.0065000006 -0.0173666680
#> [416]  0.0142666670  0.0187666660  0.0048000010 -0.0305000000  0.0237333330
#> [421] -0.0368666700  0.0124666660 -0.0063666664  0.0315333340 -0.0123333335
#> [426] -0.0118999990 -0.0517999980  0.0116000000 -0.0284333320 -0.0216666660
#> [431] -0.0088333340  0.0301333310 -0.0229333340 -0.0103666670 -0.0082333330
#> [436]  0.0553333300  0.0108333330 -0.0862333400 -0.0289666650 -0.0333333350
#> [441] -0.0115000010  0.0546000000 -0.0046333340 -0.0206666660  0.0461333300
#> [446] -0.0599333350  0.0422333330  0.0298666670 -0.0033999980  0.0301999990
#> [451]  0.0466000000  0.0823666700  0.0155333310 -0.0075333323  0.0283333300
#> [456] -0.0372000000 -0.0395333320 -0.0278000000 -0.0389333330 -0.0145999985
#> [461]  0.0181333330 -0.0361333300  0.0245333330 -0.0060000005 -0.0030999996
#> [466]  0.0254666660 -0.0329666660  0.0050333333  0.0287000000 -0.0443333350
#> [471] -0.0344333300  0.0285666680 -0.0086999990  0.0285000000 -0.0511000000
#> [476] -0.0011000000 -0.0001666670  0.0481999960  0.0195999980  0.0024666667
#> [481]  0.0274333300 -0.0226000000  0.0208333340 -0.0117333340  0.0077666673
#> [486] -0.0191999990 -0.0174000020 -0.0112666660 -0.0350000000 -0.0555000000
#> [491] -0.0201333330  0.0265333340  0.0403666650  0.0122000010 -0.0227666660
#> [496]  0.0074666664 -0.0552000000  0.0368000000 -0.0743666700 -0.0193333330
#> [501] -0.0038666672  0.0262333320 -0.0053666667  0.0255666670  0.0015666666
#> [506]  0.0105666680 -0.0344333350 -0.0185666680  0.0036666666  0.0741666600
#> [511] -0.0253999980 -0.0093333340
#> 
#>