Rosette API sentence determination service

Divides the input into sentences. The response contains a list of sentences.

ros_sentences(content, genre = NULL, language = NULL)

Arguments

content

either a URI or character vector of content to process

genre

document genre (optional)

language

Language: ISO 639-3 code (optional)

Examples

lines <- c("This land is your land. This land is my land",
           "From California to the New York island;",
           "From the red wood forest to the Gulf Stream waters",
           "",  "This land was made for you and Me.",
           "", "As I was walking that ribbon of highway,",
           "I saw above me that endless skyway:",
           "I saw below me that golden valley:",
           "This land was made for you and me.")
ros_sentences(paste0(lines, collapse="\\n"))

#> $sentences
#> [1] "This land is your land. "                                                                                                                                                                                                                                                                                                 
#> [2] "This land is my land\\nFrom California to the New York island;\\nFrom the red wood forest to the Gulf Stream waters\\n\\nThis land was made for you and Me.\\n\\nAs I was walking that ribbon of highway,\\nI saw above me that endless skyway:\\nI saw below me that golden valley:\\nThis land was made for you and me."
#> 
#>